refactor: remove advanced wallet mode
