--- src/components/CurrentWalletAdvanced.module.css
@@ -1,7 +1,0 @@
-.current-wallet-placeholder-container {
-  margin-bottom: 0.375rem;
-}
-
-.current-wallet-placeholder {
-  height: 3.3rem;
-}
