+++ src/i18n/locales/en/translation.json
@@ -148,8 +148,6 @@
     "use_btc": "Display amounts in bitcoin",
     "use_dark_theme": "Switch to dark theme",
     "use_light_theme": "Switch to light theme",
-    "use_normal_mode": "Use normal wallet mode",
-    "use_dev_mode": "Use developer wallet mode",
     "text_help_translate": "Missing your language? Help us out!",
     "show_seed": "Show seed phrase",
     "hide_seed": "Hide seed phrase",
@@ -190,9 +188,7 @@
     "placeholder_recipient": "Enter address or choose jar from wallet...",
     "feedback_reused_address": "This address is already used. To preserve your privacy please choose another one.",
     "label_account": "Send from",
-    "label_account_dev_mode": "Account to send from",
     "account_selector_option": "Jar {{ number }}",
-    "account_selector_option_dev_mode": "Account {{ number }}",
     "label_amount": "Amount in sats",
     "placeholder_amount": "Enter amount...",
     "feedback_invalid_amount": "Please provide a valid amount.",
